refactor(api): move scm logic to separate package (#848)
