+++ api/scm/doc.go
@@ -1,0 +1,10 @@
+// Copyright (c) 2023 Target Brands, Inc. All rights reserved.
+//
+// Use of this source code is governed by the LICENSE file in this repository.
+
+// Package scm provides the scm handlers for the Vela API.
+//
+// Usage:
+//
+//	import "github.com/go-vela/server/api/scm"
+package scm
